Update gemspec with modern values
Requires a major version of 1 and a minor version of >= .7.  Patch level is
not considered.

This also explicitly adds major version 2 to rspec, since that was what
was used in the original Gemfile.lock.  This allows someone to clone the
repo and immediately run `bundle exec rspec` to see test coverage.
Previosuly, an newly-cloned repo would error out since `rspec-core` was
not included.

Note that 19 of the tests are failing.  This does not seem to be not
related to the JSON or jwt versions, though - it happens in the main
repo as well as soon as one fixes the 'rspec-core not included' error.